==> AllData/2.데이터전처리/2_2.All_Data_결측치 채우기.ipynb <==
목표 : 상장기업 결측치 채우기
=====================
- inf는 max로 대체, -inf는 min으로 대체
- 정규성 검사
- 정규성 -> fillna(mean 평균)
- 정규성X -> fillna(median 중앙값)

import modules

import pandas as pd
import numpy as np

#워닝 메시지
import warnings
warnings.filterwarnings('ignore')

import os
current_directory = os.getcwd()
print("현재 디렉토리:", current_directory)

from scipy.stats import shapiro, anderson, kstest, norm

환경설정

# pd.set_option('display.max_rows', None)
pd.reset_option('display.max_rows')

# pd.set_option('display.max_columns', None)
pd.reset_option('display.max_columns')

pd.set_option('display.max_seq_items', None)

pd.set_option('float_format', '{:.3f}'.format) # 소수점 3개까지 출력

load file

all_data = pd.read_csv("../1.데이터셋/Feature/all_data_train_data.csv")
# 영업년수와 상장년수가 있더라

all_data.columns

# describe

all

# train data 결측치 개수
all_data.isna().sum().sort_values(ascending=False)

# train data 0의 개수 비율
all_data.drop(['부실판단', '도입기', '성장기', '성숙기', '쇠퇴기'], axis=1).isin([0]).sum().sort_values(ascending=False) / len(all_data) * 100

# train data inf 값 개수
all_data.isin([np.inf]).sum().sort_values(ascending=False)

# train data -inf 값 개수
all_data.isin([-np.inf]).sum().sort_values(ascending=False)

# 범주 분할 1현금흐름 / 2안정성 / 2수익성 / 3유동성 / 4성장성 / 5생산성 / 6활동성 / 7시장지표 및 기타

# 비상장 all data는 unlisted
# 비상장 label들

labels = all_data[
    [
        '거래소코드',
        '회계년도',
        '부실판단',
        '도입기',
        '성숙기',
        '성장기',
        '쇠퇴기'
    ]
]

# 비상장 피처
features = all_data.drop(
    labels=[
        '회사명',
        '거래소코드',
        '회계년도',        
        '도입기',
        '성숙기',
        '성장기',
        '쇠퇴기',
        '부실판단'        
    ],
    axis=1,    
)

# 1현금흐름 4개
cash_ft1 = all_data[
    [
        'CASH FLOW 대 부채비율',
        'CASH FLOW 대 차입금비율',
        'CASH FLOW 대 총자본비율',
        'CASH FLOW 대 매출액비율',
        '부실판단'
    ]
]

# 2안정성 9개
stability_ft1 = all_data[
    [
        '유동비율',
        '당좌비율',
        '부채비율',
        '유동부채비율',
        '차입금의존도',
        '순운전자본비율',
        '현금비율',
        '유동자산구성비율',
        '자기자본구성비율',
        '부실판단'
    ]
]

# 3수익성 8개
profitability_ft1 = all_data[
    [
        '자기자본순이익률',
        '경영자본순이익률',
        '매출액순이익률',
        '총자본사업이익률',
        '총자본영업이익률',
        '매출액영업이익률',
        '금융비용부담률',
        'CROE',
        '부실판단'
    ]
]

# 4성장성 6개
growth_ft1 = all_data[
    [
        '총자산증가율',
        '총자본증가율',
        '유동자산증가율',
        '비유동자산증가율',
        '매출액증가율',
        '유형자산증가율',
        '부실판단'
    ]
]

# 5생산성 6개
productivity_ft1 = all_data[
    [
        '부가가치',
        '총자본투자효율',
        '설비투자효율',
        '부가가치율',
        '이윤분배율',
        '자본분배율',
        '부실판단'
    ]
]

# 6활동성 8개
activity_ft1 = all_data[
    [
        '총자산회전율',
        '매출채권회전률',
        '당좌자산회전률',
        '유형자산회전율',
        '경영자본회전률',
        '총자본회전률',
        '비유동자산회전률',
        '유동자산회전률',
        '부실판단'
    ]
]

# 7기타
etc_ft1 = all_data[
    [
        '영업년수',
        '쭈피처',
        '영업투자현금비율',
        '영업재무현금비율', 
        '도입기',
        '성숙기',
        '성장기',
        '쇠퇴기',
        '부실판단'
    ]
]

# 1현금흐름 4가지 cash
cash_ft1.describe()

# 2안정성 9가지 stability
stability_ft1.describe()

# 3수익성 8가지 profitability
productivity_ft1.describe()

# 4성장성 6개 growth
growth_ft1.describe()

# 5생산성 6가지 productivity
productivity_ft1.describe()

# 6활동성 8가지 activity
activity_ft1.describe()

# 7기타 etc
etc_ft1.describe()

# 기업수명주기와 부실별

# 도입정상
도입정상 = all_data[(all_data['도입기'] == 1) & (all_data['부실판단'] == 0)]

# 도입정상 결측치확인
도입정상.isna().sum().sort_values(ascending=False)

# 도입부실
도입부실 = all_data[(all_data['도입기'] == 1) & (all_data['부실판단'] == 1)]

# 도입부실 결측치확인
도입부실.isna().sum().sort_values(ascending=False)

# 성장정상
성장정상 = all_data[(all_data['성장기'] == 1) & (all_data['부실판단'] == 0)]

# 성장정사 결측치확인
성장정상.isna().sum().sort_values(ascending=False)

# 성장부실
성장부실 = all_data[(all_data['성장기'] == 1) & (all_data['부실판단'] == 1)]

inf_col = 성장부실.isin([np.inf]).sum().sort_values(ascending=False).index[0]
max_value = 성장부실[inf_col].replace(np.inf, np.nan).max()
성장부실[inf_col] = 성장부실[inf_col].replace(np.inf, max_value)

# 성장부실 결측치 확인
성장부실.isna().sum().sort_values(ascending=False)

# 성숙정상
성숙정상 = all_data[(all_data['성숙기'] == 1) & (all_data['부실판단'] == 0)]

# 성숙정상 결측치 확인
성숙정상.isna().sum().sort_values(ascending=False)

# 성숙부실
성숙부실 = all_data[(all_data['성숙기'] == 1) & (all_data['부실판단'] == 1)]

# 성숙부실 결측치 확인
성숙부실.isna().sum().sort_values(ascending=False)

# 쇠퇴정상
쇠퇴정상 = all_data[(all_data['쇠퇴기'] == 1) & (all_data['부실판단'] == 0)]

# 쇠퇴정상 결측치 확인
쇠퇴정상.isna().sum().sort_values(ascending=False)

# 정규성 컬럼
na_col0 = 쇠퇴정상.isna().sum().sort_values(ascending=False).index[0]

# 테스트를 위한 샘플 데이터 생성
data0 = 쇠퇴정상[na_col0].dropna()

# 샤피로-윌크 검정 수행
statistic, p_value = shapiro(data0)

# 검정 통계량과 p-value 출력
print(f"검정 통계량: {statistic}, p-value: {p_value}")

# p-value를 기준으로 귀무가설을 기각 여부 판단
alpha = 0.05
if p_value > alpha:
    print("데이터는 정규 분포를 따릅니다 (귀무가설을 기각하지 않음)")
else:
    print("데이터는 정규 분포를 따르지 않습니다 (귀무가설을 기각)")

# 안더슨-달링 검정 수행
result = anderson(data0)

# 검정 통계량과 임계값 출력
print(f"검정 통계량: {result.statistic}")

# 임계값과 유의수준에 따른 결과 출력
for i in range(len(result.critical_values)):
    sl, cv = result.significance_level[i], result.critical_values[i]
    print(f"유의수준 {sl}%: 임계값 = {cv}")

# 검정 통계량이 임계값보다 작으면 정규 분포를 따른다고 판단
if result.statistic < result.critical_values[2]:
    print("데이터는 정규 분포를 따릅니다.")
else:
    print("데이터는 정규 분포를 따르지 않습니다.")

# Kolmogorov-Smirnov 검정 수행
statistic, p_value = kstest(data0, 'norm')

# 검정 통계량과 p-value 출력
print(f"검정 통계량: {statistic}, p-value: {p_value}")

# p-value를 기준으로 귀무가설을 기각 여부 판단
alpha = 0.05
if p_value > alpha:
    print("데이터는 정규 분포를 따릅니다 (귀무가설을 기각하지 않음)")
else:
    print("데이터는 정규 분포를 따르지 않습니다 (귀무가설을 기각)")

# fillna median
쇠퇴정상[na_col0].fillna(data0.median(), inplace=True)

# 쇠퇴부실
쇠퇴부실 = all_data[(all_data['쇠퇴기'] == 1) & (all_data['부실판단'] == 1)]

# 쇠퇴부실 결측치 확인
쇠퇴부실.isna().sum().sort_values(ascending=False)

# 결측치 컬럼
na_col0 = 쇠퇴부실.isna().sum().sort_values(ascending=False).index[0]

# 테스트를 위한 샘플 데이터 생성
data0 = 쇠퇴부실[na_col0].dropna()

# 샤피로-윌크 검정 수행
statistic, p_value = shapiro(data0)

# 검정 통계량과 p-value 출력
print(f"검정 통계량: {statistic}, p-value: {p_value}")

# p-value를 기준으로 귀무가설을 기각 여부 판단
alpha = 0.05
if p_value > alpha:
    print("데이터는 정규 분포를 따릅니다 (귀무가설을 기각하지 않음)")
else:
    print("데이터는 정규 분포를 따르지 않습니다 (귀무가설을 기각)")

# 안더슨-달링 검정 수행
result = anderson(data0)

# 검정 통계량과 임계값 출력
print(f"검정 통계량: {result.statistic}")

# 임계값과 유의수준에 따른 결과 출력
for i in range(len(result.critical_values)):
    sl, cv = result.significance_level[i], result.critical_values[i]
    print(f"유의수준 {sl}%: 임계값 = {cv}")

# 검정 통계량이 임계값보다 작으면 정규 분포를 따른다고 판단
if result.statistic < result.critical_values[2]:
    print("데이터는 정규 분포를 따릅니다.")
else:
    print("데이터는 정규 분포를 따르지 않습니다.")

# Kolmogorov-Smirnov 검정 수행
statistic, p_value = kstest(data0, 'norm')

# 검정 통계량과 p-value 출력
print(f"검정 통계량: {statistic}, p-value: {p_value}")

# p-value를 기준으로 귀무가설을 기각 여부 판단
alpha = 0.05
if p_value > alpha:
    print("데이터는 정규 분포를 따릅니다 (귀무가설을 기각하지 않음)")
else:
    print("데이터는 정규 분포를 따르지 않습니다 (귀무가설을 기각)")

# fillna(median)
쇠퇴부실[na_col0].fillna(data0.median(), inplace=True)

# to_csv

all_data.loc[도입부실.index] = 도입부실
all_data.loc[도입정상.index] = 도입정상
all_data.loc[성장부실.index] = 성장부실
all_data.loc[성장정상.index] = 성장정상
all_data.loc[성숙부실.index] = 성숙부실
all_data.loc[성숙정상.index] = 성숙정상
all_data.loc[쇠퇴부실.index] = 쇠퇴부실
all_data.loc[쇠퇴정상.index] = 쇠퇴정상

all_data.isna().sum()

all_data.to_csv('../1.데이터셋/Feature/all_data_filled_train_data.csv', index=False)

all_data.isin([np.inf]).sum()
